refactor(codegen): extract getCommandProperties to parsers-common
